working on the following feature #108 and TravisCI #153

+++ server/routes/timeline.js
@@ -27,7 +27,7 @@ router1.post('/timeline', (req, res) => {
     }
     else {
  //if it has followings: then we will display the wits of his followings and his.
-      if (row[0].following > 0) {
+      if (row[0].following >= 0) {
         sqlQuery1 = "SELECT * FROM events WHERE username IN (SELECT follow_name AND username FROM following WHERE username = ?)"
         connection.connection.query(sqlQuery1, userLoggedIN, function (err, results) {
           if (err) {
@@ -100,6 +100,10 @@ router1.get('/likedWits', (req, res) => {
 router1.post('/like', (req, res) => {
 //we will get the wit_id from the frontend:
   witInfo = req.body;
+  if(userLoggedIN === witInfo.username && witInfo.username !=null){
+    res.status(401).json("user can't like their own wit");
+    return;
+  }
 //updating the table of events by increasing the likes number of this wit:
   sqlQuery2 = "UPDATE events SET numOfLikes = numOfLikes + 1, boolvalue = true WHERE wit_id = ? ";
   connection.connection.query(sqlQuery2, witInfo.wit_id, function (err, result) {
@@ -180,6 +184,10 @@ router1.post('/likesList', function (req, res) {
 
 router1.post('/deleteWit', (req, res) => {
   witInfo = req.body;
+  if(witInfo.username != userLoggedIN && witInfo.username !=null){
+    res.status(401).json("user can't delete others' wits");
+    return;
+  }
 //Decreasing the likes number in the events table related to this wit:
   sqlQueryDelete = "DELETE FROM events WHERE wit_id = ?";
   connection.connection.query(sqlQueryDelete, witInfo.wit_id, function (err, result) {
@@ -203,6 +211,10 @@ router1.post('/witPost', (req, res) => {
       numOfLikes: 0,
       numOfReplies: 0
   }
+if (postInfo.wit.length > 120){
+  res.status(401).json("Too long wit");
+      return;
+}
 if (postInfo.wit.length == 0) {
   res.status(401).json("Empty wit");
       return;
